Problem: No way to deliver DoAuth client
Solution: After some deliberation we figured that putting an in-browser
JS file into an NPM package to let UNPKG distribute it is a way to go.

We're still figuring out how to make sure that sodium-wrapper is
properly bundled, but we will likely simply detect its absence in the
set-up function and print an error with instructions into the console.